beatbuddy/facial2/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
Automat==22.10.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.8
cssselect==1.2.0
cycler==0.12.1
deepface==0.0.93
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.6
et-xmlfile==1.1.0
fake-useragent==1.5.1
filelock==3.14.0
fire==0.7.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.68.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itemadapter==0.9.0
itemloaders==1.3.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
keras==3.7.0
kiwisolver==1.4.5
libclang==18.1.1
lxml==5.2.2
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.4.1
mtcnn==1.0.0
multidict==6.0.5
namex==0.0.8
numpy==2.0.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parsel==1.9.1
pillow==10.3.0
platformdirs==4.2.1
Protego==0.3.1
protobuf==5.29.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
Pygments==2.18.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyVirtualDisplay==3.0
PyYAML==6.0.1
queuelib==1.7.0
redis==5.2.0
requests==2.32.3
requests-file==2.1.0
retina-face==0.0.17
rich==13.9.4
scikit-learn==1.5.0
scipy==1.14.0
Scrapy==2.11.2
selectorlib==0.16.0
selenium==4.22.0
service-identity==24.1.0
setuptools==70.3.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spotipy==2.24.0
SQLAlchemy==2.0.36
sqlparse==0.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tldextract==5.1.2
tqdm==4.67.1
trio==0.26.0
trio-websocket==0.11.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.1
w3lib==2.2.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.45.1
wrapt==1.17.0
wsproto==1.2.0
yarl==1.9.4
zope.interface==6.4.post2
from flask_mail import Mail, Message